Rename Hash object to Dictionary
Hash is too vague of a term (and more an impl. detail) for this form of object (although Ruby does call them this..). The correct term would be Map, however this could conflict with FP principles within the language. As such, I have opted to call them a Dictionary similiar to how Python does.